requirements.txt
streamlit==1.40.0
google-genai==0.5.1
python-dotenv==1.0.0
requests==2.32.3
